+++ tests/array/utils/gtest1dTest.h
@@ -6,14 +6,14 @@ extern "C" {
 
 #define SCALAR double
 
-const uint ARRAY_SIZE = 11;
+const size_t ARRAY_SIZE = 11;
 
 class Array1dTest : public ::testing::Test {
 public:
-  uint IterAbsOffset(array1d::iterator iter) {
+  size_t IterAbsOffset(array1d::iterator iter) {
     return iter.i();
   }
-  uint IterAbsOffset(array1d::const_iterator citer) {
+  size_t IterAbsOffset(array1d::const_iterator citer) {
     return citer.i();
   }
 
@@ -35,7 +35,7 @@ protected:
   static array1d::const_pointer cptr, cptr2;
   static array1d::iterator iter, iter2;
   static array1d::const_iterator citer, citer2;
-  static uint offset, viewLen;
+  static size_t offset, viewLen;
 };
 
 array1d Array1dTest::arr(ARRAY_SIZE, ZFP_RATE_PARAM_BITS);
@@ -44,4 +44,4 @@ array1d::pointer Array1dTest::ptr, Array1dTest::ptr2;
 array1d::const_pointer Array1dTest::cptr, Array1dTest::cptr2;
 array1d::iterator Array1dTest::iter, Array1dTest::iter2;
 array1d::const_iterator Array1dTest::citer, Array1dTest::citer2;
-uint Array1dTest::offset, Array1dTest::viewLen;
+size_t Array1dTest::offset, Array1dTest::viewLen;
