+++ tests/array/utils/gtest2dTest.h
@@ -6,15 +6,15 @@ extern "C" {
 
 #define SCALAR double
 
-const uint ARRAY_SIZE_X = 11;
-const uint ARRAY_SIZE_Y = 5;
+const size_t ARRAY_SIZE_X = 11;
+const size_t ARRAY_SIZE_Y = 5;
 
 class Array2dTest : public ::testing::Test {
 public:
-  uint IterAbsOffset(array2d::iterator iter) {
+  size_t IterAbsOffset(array2d::iterator iter) {
     return iter.i() + ARRAY_SIZE_X * iter.j();
   }
-  uint IterAbsOffset(array2d::const_iterator citer) {
+  size_t IterAbsOffset(array2d::const_iterator citer) {
     return citer.i() + ARRAY_SIZE_X * citer.j();
   }
 
@@ -40,7 +40,7 @@ protected:
   static array2d::const_pointer cptr, cptr2;
   static array2d::iterator iter, iter2;
   static array2d::const_iterator citer, citer2;
-  static uint offsetX, offsetY, viewLenX, viewLenY;
+  static size_t offsetX, offsetY, viewLenX, viewLenY;
 };
 
 array2d Array2dTest::arr(ARRAY_SIZE_X, ARRAY_SIZE_Y, ZFP_RATE_PARAM_BITS);
@@ -49,4 +49,4 @@ array2d::pointer Array2dTest::ptr, Array2dTest::ptr2;
 array2d::const_pointer Array2dTest::cptr, Array2dTest::cptr2;
 array2d::iterator Array2dTest::iter, Array2dTest::iter2;
 array2d::const_iterator Array2dTest::citer, Array2dTest::citer2;
-uint Array2dTest::offsetX, Array2dTest::offsetY, Array2dTest::viewLenX, Array2dTest::viewLenY;
+size_t Array2dTest::offsetX, Array2dTest::offsetY, Array2dTest::viewLenX, Array2dTest::viewLenY;
