+++ tests/array/utils/gtest2fTest.h
@@ -6,15 +6,15 @@ extern "C" {
 
 #define SCALAR float
 
-const uint ARRAY_SIZE_X = 11;
-const uint ARRAY_SIZE_Y = 5;
+const size_t ARRAY_SIZE_X = 11;
+const size_t ARRAY_SIZE_Y = 5;
 
 class Array2fTest : public ::testing::Test {
 public:
-  uint IterAbsOffset(array2f::iterator iter) {
+  size_t IterAbsOffset(array2f::iterator iter) {
     return iter.i() + ARRAY_SIZE_X * iter.j();
   }
-  uint IterAbsOffset(array2f::const_iterator citer) {
+  size_t IterAbsOffset(array2f::const_iterator citer) {
     return citer.i() + ARRAY_SIZE_X * citer.j();
   }
 
@@ -40,7 +40,7 @@ protected:
   static array2f::const_pointer cptr, cptr2;
   static array2f::iterator iter, iter2;
   static array2f::const_iterator citer, citer2;
-  static uint offsetX, offsetY, viewLenX, viewLenY;
+  static size_t offsetX, offsetY, viewLenX, viewLenY;
 };
 
 array2f Array2fTest::arr(ARRAY_SIZE_X, ARRAY_SIZE_Y, ZFP_RATE_PARAM_BITS);
@@ -49,4 +49,4 @@ array2f::pointer Array2fTest::ptr, Array2fTest::ptr2;
 array2f::const_pointer Array2fTest::cptr, Array2fTest::cptr2;
 array2f::iterator Array2fTest::iter, Array2fTest::iter2;
 array2f::const_iterator Array2fTest::citer, Array2fTest::citer2;
-uint Array2fTest::offsetX, Array2fTest::offsetY, Array2fTest::viewLenX, Array2fTest::viewLenY;
+size_t Array2fTest::offsetX, Array2fTest::offsetY, Array2fTest::viewLenX, Array2fTest::viewLenY;
