+++ tests/array/utils/gtest3fTest.h
@@ -6,16 +6,16 @@ extern "C" {
 
 #define SCALAR float
 
-const uint ARRAY_SIZE_X = 11;
-const uint ARRAY_SIZE_Y = 18;
-const uint ARRAY_SIZE_Z = 5;
+const size_t ARRAY_SIZE_X = 11;
+const size_t ARRAY_SIZE_Y = 18;
+const size_t ARRAY_SIZE_Z = 5;
 
 class Array3fTest : public ::testing::Test {
 public:
-  uint IterAbsOffset(array3f::iterator iter) {
+  size_t IterAbsOffset(array3f::iterator iter) {
     return iter.i() + ARRAY_SIZE_X * iter.j() + ARRAY_SIZE_X * ARRAY_SIZE_Y * iter.k();
   }
-  uint IterAbsOffset(array3f::const_iterator citer) {
+  size_t IterAbsOffset(array3f::const_iterator citer) {
     return citer.i() + ARRAY_SIZE_X * citer.j() + ARRAY_SIZE_X * ARRAY_SIZE_Y * citer.k();
   }
 
@@ -45,8 +45,8 @@ protected:
   static array3f::const_pointer cptr, cptr2;
   static array3f::iterator iter, iter2;
   static array3f::const_iterator citer, citer2;
-  static uint offsetX, offsetY, offsetZ;
-  static uint viewLenX, viewLenY, viewLenZ;
+  static size_t offsetX, offsetY, offsetZ;
+  static size_t viewLenX, viewLenY, viewLenZ;
 };
 
 array3f Array3fTest::arr(ARRAY_SIZE_X, ARRAY_SIZE_Y, ARRAY_SIZE_Z, ZFP_RATE_PARAM_BITS);
@@ -55,5 +55,5 @@ array3f::pointer Array3fTest::ptr, Array3fTest::ptr2;
 array3f::const_pointer Array3fTest::cptr, Array3fTest::cptr2;
 array3f::iterator Array3fTest::iter, Array3fTest::iter2;
 array3f::const_iterator Array3fTest::citer, Array3fTest::citer2;
-uint Array3fTest::offsetX, Array3fTest::offsetY, Array3fTest::offsetZ;
-uint Array3fTest::viewLenX, Array3fTest::viewLenY, Array3fTest::viewLenZ;
+size_t Array3fTest::offsetX, Array3fTest::offsetY, Array3fTest::offsetZ;
+size_t Array3fTest::viewLenX, Array3fTest::viewLenY, Array3fTest::viewLenZ;
