+++ tests/array/utils/gtest4fTest.h
@@ -6,17 +6,17 @@ extern "C" {
 
 #define SCALAR float
 
-const uint ARRAY_SIZE_X = 14;
-const uint ARRAY_SIZE_Y = 9;
-const uint ARRAY_SIZE_Z = 7;
-const uint ARRAY_SIZE_W = 6;
+const size_t ARRAY_SIZE_X = 14;
+const size_t ARRAY_SIZE_Y = 9;
+const size_t ARRAY_SIZE_Z = 7;
+const size_t ARRAY_SIZE_W = 6;
 
 class Array4fTest : public ::testing::Test {
 public:
-  uint IterAbsOffset(array4f::iterator iter) {
+  size_t IterAbsOffset(array4f::iterator iter) {
     return iter.i() + ARRAY_SIZE_X * iter.j() + ARRAY_SIZE_X * ARRAY_SIZE_Y * iter.k() + ARRAY_SIZE_X * ARRAY_SIZE_Y * ARRAY_SIZE_Z * iter.l();
   }
-  uint IterAbsOffset(array4f::const_iterator citer) {
+  size_t IterAbsOffset(array4f::const_iterator citer) {
     return citer.i() + ARRAY_SIZE_X * citer.j() + ARRAY_SIZE_X * ARRAY_SIZE_Y * citer.k() + ARRAY_SIZE_X * ARRAY_SIZE_Y * ARRAY_SIZE_Z * citer.l();
   }
 
@@ -50,8 +50,8 @@ protected:
   static array4f::const_pointer cptr, cptr2;
   static array4f::iterator iter, iter2;
   static array4f::const_iterator citer, citer2;
-  static uint offsetX, offsetY, offsetZ, offsetW;
-  static uint viewLenX, viewLenY, viewLenZ, viewLenW;
+  static size_t offsetX, offsetY, offsetZ, offsetW;
+  static size_t viewLenX, viewLenY, viewLenZ, viewLenW;
 };
 
 array4f Array4fTest::arr(ARRAY_SIZE_X, ARRAY_SIZE_Y, ARRAY_SIZE_Z, ARRAY_SIZE_W, ZFP_RATE_PARAM_BITS);
@@ -60,5 +60,5 @@ array4f::pointer Array4fTest::ptr, Array4fTest::ptr2;
 array4f::const_pointer Array4fTest::cptr, Array4fTest::cptr2;
 array4f::iterator Array4fTest::iter, Array4fTest::iter2;
 array4f::const_iterator Array4fTest::citer, Array4fTest::citer2;
-uint Array4fTest::offsetX, Array4fTest::offsetY, Array4fTest::offsetZ, Array4fTest::offsetW;
-uint Array4fTest::viewLenX, Array4fTest::viewLenY, Array4fTest::viewLenZ, Array4fTest::viewLenW;
+size_t Array4fTest::offsetX, Array4fTest::offsetY, Array4fTest::offsetZ, Array4fTest::offsetW;
+size_t Array4fTest::viewLenX, Array4fTest::viewLenY, Array4fTest::viewLenZ, Array4fTest::viewLenW;
